requirements.txt
# FRbox - Face Recognition Service Dependencies
# Python 3.11

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Face Recognition
face_recognition>=1.3.0
Pillow>=10.2.0

# Numerical computing
numpy>=1.26.3

# Pydantic for validation
pydantic>=2.5.3

# HTTP client (for testing)
requests>=2.31.0

# Python standard library
python-multipart>=0.0.6
